refactor(checker): split explicit any annotation helpers
